% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-BipartiteSBM.R
\name{BipartiteSBM_sampler}
\alias{BipartiteSBM_sampler}
\title{R6 class for Simple SBM sampler}
\description{
R6 class for Simple SBM sampler

R6 class for Simple SBM sampler
}
\section{Super class}{
\code{\link[GSBM:SBM]{GSBM::SBM}} -> \code{BipartiteSBM_sampler}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-rBlocks}{\code{BipartiteSBM_sampler$rBlocks()}}
\item \href{#method-clone}{\code{BipartiteSBM_sampler$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="GSBM" data-topic="SBM" data-id="initialize">}\href{../../GSBM/html/SBM.html#method-initialize}{\code{GSBM::SBM$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rBlocks"></a>}}
\if{latex}{\out{\hypertarget{method-rBlocks}{}}}
\subsection{Method \code{rBlocks()}}{
a method to generate a vector of clusters indicators
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BipartiteSBM_sampler$rBlocks()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BipartiteSBM_sampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
