% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-SBM.R
\name{coef.SBM}
\alias{coef.SBM}
\title{Extract model coefficients}
\usage{
\method{coef}{SBM}(object, type = c("connectivity", "block", "covariates"), ...)
}
\arguments{
\item{object}{an R6 object inheriting from class SBM_fit (like SimpleSBM_fit or BipartiteSBM_fit)}

\item{type}{type of parameter that should be extracted. Either 'block' for \deqn{\pi}, 'connectivity' for \deqn{\theta},
or "covariates" for \deqn{\beta}. Default is 'connectivity'.}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
vector or list of parameters.
}
\description{
Extracts model coefficients from objects with class \code{\link[=SBM]{SBM}} and children (\code{\link[=SimpleSBM_fit]{SimpleSBM_fit}},
\code{\link[=BipartiteSBM_fit]{BipartiteSBM_fit}})
}
