% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-ScoreSBM.R
\name{coef.ScoreSBM}
\alias{coef.ScoreSBM}
\title{Extract model coefficients}
\usage{
\method{coef}{ScoreSBM}(object, type = c("connectivity", "membership", "emission"), ...)
}
\arguments{
\item{object}{an R6 object inheriting from class ScoreSBM_fit}

\item{type}{type of parameter that should be extracted. Either 'memberships' for \deqn{\pi}, 'connectivity' for \deqn{\alpha},
or "emission" for \deqn{\mu_0,\Sigma_0, \mu_1, \sigma_1}.}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
vector or list of parameters.
}
\description{
Extracts model coefficients from objects with class \code{\link[=ScoreSBM]{ScoreSBM}}
}
