% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimateMultipartiteSBM}
\alias{estimateMultipartiteSBM}
\title{Estimation for multipartite SBM}
\usage{
estimateMultipartiteSBM(listSBM, estimOptions = list())
}
\arguments{
\item{estimOptions}{options for the inference procedure}

\item{ldefinedNet}{list of networks that were defined by the \code{defineSBM} function}
}
\value{

}
\description{
Estimation for multipartite SBM
}
