% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimateMultiplexSBM}
\alias{estimateMultiplexSBM}
\title{Estimation for Multiplex SBM}
\usage{
estimateMultiplexSBM(listSBM, dependent = FALSE, estimOptions = list())
}
\arguments{
\item{listSBM}{list of networks that were defined by the \code{defineSBM} function}

\item{dependent}{logical parameter indicating whether the networks in the multiplex structure are dependent beyond the latent variables,}

\item{estimOptions}{options for the inference procedure}
}
\value{
a MultiplexSBM_fit object with the estimated parameters and the blocks
}
\description{
Estimation for Multiplex SBM
}
\details{
The list of parameters \code{estimOptions} essentially tunes the optimization process and the variational EM algorithm. See the details in the function "estimateMultipartiteSBM" if dependent=FALSE, details in the function "estimateSimpleSBM" otherwise.
}
\examples{
## MultiplexSBM with no dependence
#TODO once the SampleMultiplexSBM  function works

## MultiplexSBM with dependence
}
