% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-SBM.R
\name{fitted.SBM}
\alias{fitted.SBM}
\title{Extract model fitted values}
\usage{
\method{fitted}{SBM}(object, ...)
}
\arguments{
\item{object}{an R6 object inheriting from SimpleSBM_fit,  BipartiteSBM_fit or MultipartiteSBM_fit}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
a matrix of expected fitted values for each dyad
}
\description{
Extracts fitted values for object with class (\code{\link[=SimpleSBM_fit]{SimpleSBM_fit}},
\code{\link[=BipartiteSBM_fit]{BipartiteSBM_fit}}) or \code{\link[=MultipartiteSBM_fit]{multipartitepartiteSBM_fit}})
}
