% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-SBM_fit.R
\name{fitted.SBM_fit}
\alias{fitted.SBM_fit}
\title{Extract model fitted values
Extracts fitted values for object with class \code{\link[=SBM_fit]{SBM_fit}} and children (\code{\link[=SimpleSBM_fit]{SimpleSBM_fit}},
\code{\link[=BipartiteSBM_fit]{BipartiteSBM_fit}})}
\usage{
\method{fitted}{SBM_fit}(object, ...)
}
\arguments{
\item{object}{an R6 object inheriting from class SBM_fit (like SimpleSBM_fit or BipartiteSBM_fit)}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
a matrix of expected fitted values for each dyad
}
\description{
Extract model fitted values
Extracts fitted values for object with class \code{\link[=SBM_fit]{SBM_fit}} and children (\code{\link[=SimpleSBM_fit]{SimpleSBM_fit}},
\code{\link[=BipartiteSBM_fit]{BipartiteSBM_fit}})
}
