% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreSBM-optimize.R
\name{initInferenceScoreSBM}
\alias{initInferenceScoreSBM}
\title{Initialization of the inference procedure
This function initialises the inference method by mixing a Gaussian mixture on the scores of each pair of nodes and a SBM on the resulting estimated network G}
\usage{
initInferenceScoreSBM(scoreList, directed = FALSE, estimOptions = list())
}
\arguments{
\item{scoreList}{a list of the Scores (matrices of size nbNodes x nbNodes)}

\item{directed}{a logical : TRUE if the underlying network is directed,  FALSE otherwise (default value FALSE).}

\item{estimOptions}{a list of parameters controlling the initialisation step of the inference method. See details.}
}
\description{
Initialization of the inference procedure
This function initialises the inference method by mixing a Gaussian mixture on the scores of each pair of nodes and a SBM on the resulting estimated network G
}
