% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-MultipartiteSBM.R
\name{is_MultipartiteSBM}
\alias{is_MultipartiteSBM}
\title{Check  if an object is MultipartiteSBM}
\usage{
is_MultipartiteSBM(Robject)
}
\arguments{
\item{Robject}{an R6 object inheriting from class MultipartiteSBM}
}
\value{
TRUE or FALSE
}
\description{
Auxiliary function to check the given class of an object
}
