% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-SBM.R
\name{is_SBM}
\alias{is_SBM}
\title{Auxiliary function to check the given class of an object}
\usage{
is_SBM(Robject)
}
\arguments{
\item{Robject}{an R6 object inheriting from class SBM}
}
\value{
TRUE or FALSE
}
\description{
Auxiliary function to check the given class of an object
}
