% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-MultipartiteSBM_sampler.R
\name{MultipartiteSBM_sampler}
\alias{MultipartiteSBM_sampler}
\title{R6 Class definition of a Multipartite SBM sampler}
\description{
R6 Class definition of a Multipartite SBM sampler

R6 Class definition of a Multipartite SBM sampler
}
\details{
This class is designed to give a representation and sample from a Multipartite SBM
}
\section{Super class}{
\code{\link[sbm:MultipartiteSBM]{sbm::MultipartiteSBM}} -> \code{MultipartiteSBM_sampler}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{indMemberships}}{matrix for clustering memberships}

\item{\code{expectation}}{expected values of connection under the current model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-show}{\code{MultipartiteSBM_sampler$show()}}
\item \href{#method-clone}{\code{MultipartiteSBM_sampler$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="MultipartiteSBM" data-id="initialize">}\href{../../sbm/html/MultipartiteSBM.html#method-initialize}{\code{sbm::MultipartiteSBM$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="MultipartiteSBM" data-id="plot">}\href{../../sbm/html/MultipartiteSBM.html#method-plot}{\code{sbm::MultipartiteSBM$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="MultipartiteSBM" data-id="print">}\href{../../sbm/html/MultipartiteSBM.html#method-print}{\code{sbm::MultipartiteSBM$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\if{latex}{\out{\hypertarget{method-show}{}}}
\subsection{Method \code{show()}}{
print/show method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultipartiteSBM_sampler$show(type = "Sampler for a Stochastic Block Model")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{character to tune the displayed name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultipartiteSBM_sampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
