% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-MultiplexSBM.R
\name{MultiplexSBM}
\alias{MultiplexSBM}
\title{R6 Class definition of a Multiplex SBM}
\description{
R6 Class definition of a Multiplex SBM

R6 Class definition of a Multiplex SBM
}
\details{
R6 virtual class for Multiplex SBM representation
}
\section{Super class}{
\code{\link[sbm:MultipartiteSBM]{sbm::MultipartiteSBM}} -> \code{MultiplexSBM}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MultiplexSBM$new()}}
\item \href{#method-clone}{\code{MultiplexSBM$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="MultipartiteSBM" data-id="plot">}\href{../../sbm/html/MultipartiteSBM.html#method-plot}{\code{sbm::MultipartiteSBM$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="MultipartiteSBM" data-id="print">}\href{../../sbm/html/MultipartiteSBM.html#method-print}{\code{sbm::MultipartiteSBM$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="MultipartiteSBM" data-id="show">}\href{../../sbm/html/MultipartiteSBM.html#method-show}{\code{sbm::MultipartiteSBM$show()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor for Multipartite SBM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiplexSBM$new(listSBM, memberships = NULL, dep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{listSBM}}{list of SimpleSBM or BipartiteSBM}

\item{\code{memberships}}{list of memberships for each node in each function group.Default value is NULL}

\item{\code{dep}}{boolean indicating whether dependence is assumed between networks beyond the common dependence on the latent variables}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiplexSBM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
