% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreSBM-optimize.R
\name{optimizeScoreSBM}
\alias{optimizeScoreSBM}
\title{Inference of the Score SBM  Model
\code{estimateScoreSBM} performs the estimation and model selection for the ScoreSBM model.}
\usage{
optimizeScoreSBM(
  scores,
  directed = FALSE,
  estimOptions = list(),
  monitoring = list()
)
}
\arguments{
\item{directed}{:  if true the inference network is directed. Default value = FALSE.}

\item{estimOptions}{: tunes the optimization process (see details below)}

\item{monitoring}{: specifies if the lowerBound along the VEM iterations is saved (monitoring = list(lowerBound = TRUE))}

\item{scoreList}{:  list of Scores for each dyad  of an underlying network}
}
\value{
The output is a list of estimated models, each one corresponding to a number of blocks.
}
\description{
Inference of the Score SBM  Model
\code{estimateScoreSBM} performs the estimation and model selection for the ScoreSBM model.
}
\details{
The list of parameters \code{estimOptions} essentially tunes the optimization process and the variational EM algorithm, with the following parameters
\itemize{
\item{"verbosity"}{ controls the verbosity of the procedure (0 or 1). Default is 1.}
\item{"exploreFactor"}{ controls the exploration of the number of groups in the initialization step, Default is 1.5}
\item{"nbBlocksRange"}{ minimal and maximal number or blocks explored. Default is c(1,Inf)}
\item{"nbCores"}{ integer for number of cores used. Default is 1. }
\item{"maxIterVE"}{ Maximum number of iterations in the VE step. Default is 100.}
\item{"maxIterVE"}{ Maximum number of iterations in the VEM algorithm. Default is 1000.}
\item{"tauTol"}{ Tolerance in the VEM algorithm. Default value \code{tauTol = 2 * .Machine$double.eps}}
\item{"etaTol"}{ Tolerance in the VEM algorithm. Default value \code{etaTol = 2 * .Machine$double.eps}}
\item{"valCritStop"}{ Algorithm stops when the difference on successive tau is below valCritStop. Default value is  1e-6}
}

Each element of the output list contains the following quantites:
\itemize{
\item{"theta"}{ estimated parameters}
\item{"nbBlocks"}{ number of blocks in the underlying network}
\item{"qDist"}{ variational approximation of the the conditional distribution q(G,Z | data)}
\item{"ICL": }{Integrated Likelihood Criterion}
\item{"pen": }{Penalty term for model selection (included in the ICL)}
\item{"lowerBound": }{Lowerbound along the VEM iterations (provided if \code{monitoring$lowerBound = TRUE})}
}
}
\examples{
nbNodes  <- 60
directed <- FALSE
blockProp <- c(1/3,1/2,1/6)
nbBlocks   <- length(blockProp)
connectParam <- matrix(rbeta(nbBlocks^2,1.5,1.5 ),nbBlocks,nbBlocks)
connectParam <- 0.5*(connectParam + t(connectParam))
emissionParam <- list()
nbScores <- 4
emissionParam$noEdgeParam <- list(mean=rep(0,nbScores));
emissionParam$noEdgeParam$var <- diag(0.1,nrow = nbScores,ncol = nbScores)
emissionParam$EdgeParam <- list( mean= 1:nbScores)
emissionParam$EdgeParam$var <-  diag(0.1,nrow = nbScores,ncol = nbScores)
dataSim <- rScoreSBM(nbNodes,directed = TRUE, blockProp,connectParam,emissionParam,seed = NULL)
scoreList <- dataSim$ScoreNetworks
resEstim <- estimateScoreSBM(scoreList,directed)
}
