% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-ScoreSBM.R
\name{plot.ScoreSBM}
\alias{plot.ScoreSBM}
\title{ScoreSBM Plot}
\usage{
\method{plot}{ScoreSBM}(
  x,
  type = c("data", "predictedNetwork"),
  ordered = TRUE,
  color = TRUE,
  ...
)
}
\arguments{
\item{x}{an R6 object inheriting from class SBM_fit (like SimpleSBM_fit or BipartiteSBM_fit)}

\item{type}{character for the type of plot: either 'data' (true connection) or 'estimatedNetwork' (fitted connection). Default to 'data'.}

\item{ordered}{logical: should the rows and columns be reordered according to the clustering? Default to \code{TRUE}.}

\item{color}{logical. Adapt colormap to the clustering. Default to \code{TRUE}.}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\description{
Basic matrix plot method for ScoreSBM object
}
