% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlluvial.R
\name{plotAlluvial}
\alias{plotAlluvial}
\title{Plot an alluvial plot between  clusterings}
\usage{
plotAlluvial(listMemberships, plotOptions = list())
}
\arguments{
\item{plotOptions}{: a list containing the options for Alluvial plots}

\item{list}{: a list vectors containing the memberships}
}
\value{
display the alluvial plot, return the plotOptions
}
\description{
Plot an alluvial plot between  clusterings
}
\examples{
listMemberships <- list(C1 = rep(c('A','B','C'),each=10),C2 = rep(c(1,2,4),10))
plotAlluvial(listMemberships)
}
