% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMyMatrix.R
\name{plotMyMatrix}
\alias{plotMyMatrix}
\title{Plot an adjacency or incidence Matrix}
\usage{
plotMyMatrix(Mat, rowLabel = NULL, colLabel = NULL)
}
\arguments{
\item{Mat}{: a matrix representing the network}

\item{rowLabel}{character : type of nodes in rows (functional group) (Default is \code{NULL})}

\item{colLabel}{character :  type of nodes in columns (functional group) (Default is \code{NULL})}
}
\value{
a ggplot object corresponding to the plot
}
\description{
Plot an adjacency or incidence Matrix
}
\examples{
M <-  matrix(sample(90,c(0,1),replace=TRUE),10,9)
plot.MyMatrix(M)
}
