% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMyMultipartite.R
\name{plotMyMultipartiteMatrix}
\alias{plotMyMultipartiteMatrix}
\title{Plot the matrices corresponding to a Multipartite Network}
\usage{
plotMyMultipartiteMatrix(listSBM, normalized = FALSE, memberships = NULL)
}
\arguments{
\item{listSBM}{: a list of objects representing the multipartite network (see)}

\item{normalized}{TRUE if the various matrices are presented in the same scale (between O and 1). FALSE otherwise. Default value FALSE}

\item{memberships}{: a list of length number of Functional Groups with clusterings inside these FG}
}
\value{
a ggplot object corresponding to the plot
}
\description{
Plot the matrices corresponding to a Multipartite Network
}
\examples{
data("multipartiteEcologicalNetwork")
Net <- multipartiteEcologicalNetwork
type='bipartite'
model = 'bernoulli'
directed = FALSE
PlantFlovis = defineSBM(Net$Inc_plant_flovis, model,type,directed,
                        dimLabels = list(row="Plants",col="Flovis"))
PlantAnt = defineSBM(Net$Inc_plant_ant,model,type,directed,
                     dimLabels =list(row = "Plants", col = "Ants"))
PlantBird = defineSBM(Net$Inc_plant_bird,model,type,directed,
                      dimLabels =list(row = "Plants",col = "Birds"))
plotMyMultipartiteMatrix(list(PlantFlovis,PlantAnt,PlantBird))


}
