% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-MultipartiteSBM.R
\name{predict.MultipartiteSBM}
\alias{predict.MultipartiteSBM}
\title{Model Predictions}
\usage{
\method{predict}{MultipartiteSBM}(object, ...)
}
\arguments{
\item{object}{an R6 object inheriting from class SBM_fit (like SimpleSBM_fit or BipartiteSBM_fit)}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
a list of matrices of expected values for each dyad
}
\description{
Make predictions from an Multipartite SBM.
}
