% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-SBM.R
\name{predict.SBM}
\alias{predict.SBM}
\title{Model Predictions}
\usage{
\method{predict}{SBM}(object, covarList = object$covarList, ...)
}
\arguments{
\item{object}{an R6 object inheriting from class SBM_fit (like SimpleSBM_fit or BipartiteSBM_fit)}

\item{covarList}{a list of covariates. By default, we use the covariates associated with the model.}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
a matrix of expected values for each dyad
}
\description{
Make predictions from an SBM.
}
