% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-ScoreSBM.R
\name{predict.ScoreSBM}
\alias{predict.ScoreSBM}
\title{Model Predictions}
\usage{
\method{predict}{ScoreSBM}(object, ...)
}
\arguments{
\item{object}{an R6 object inheriting from class ScoreSBM_fit}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
a matrix of expected values for each dyad of the underlying network G
}
\description{
Make predictions from a Score  SBM.
}
