% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sampleMultiplexSBM}
\alias{sampleMultiplexSBM}
\title{Sampling of Multiplex SBMs}
\usage{
sampleMultiplexSBM(
  nbNodes,
  blockProp,
  nbLayers,
  connectParam,
  model,
  type = c("directed", "undirected", "bipartite"),
  dependent = FALSE,
  dimLabels = NULL,
  seed = NULL
)
}
\arguments{
\item{nbNodes}{number of nodes in each functional group involved in the Multiplex network}

\item{blockProp}{a vector for block proportion if the networks are simple, a list of parameters for block proportions for both functional groups if the networks are bipartite}

\item{nbLayers}{a matrix with two columns and nbNetworks lines, each line specifying the index of the functional groups in interaction.}

\item{connectParam}{list of parameters for connectivity (of length nbNetworks). Each element is a list of one or two elements: a matrix of means 'mean' and an optional matrix of variances 'var', the sizes of which must match \code{blockProp} length}

\item{model}{a vector of characters describing the model for  each network of the Multiplex relation between nodes (\code{'bernoulli'}, \code{'poisson'}, \code{'gaussian'}, ...). Default is \code{'bernoulli'}.}

\item{type}{a string of character indicating whether the networks are directed, undirected or bipartite}

\item{dependent}{connection parameters in each network}

\item{dimLabels}{an optional list of labels for functional group involved in the network}

\item{seed}{numeric to set the seed.}
}
\value{
a list of two elements : \code{simulatedMemberships} are the clustering of each node in each Functional Group,  \code{MultiplexNetwork} is the list of the simulated networks (each one being  a simple or bipartite network)
}
\description{
This function samples a Multiplex Stochastic Block Models, with various model
for the distribution of the edges:  Bernoulli, Poisson, or Gaussian models
}
\examples{
### =======================================

}
