% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm-package.R
\docType{package}
\name{sbm-package}
\alias{sbm}
\alias{sbm-package}
\title{sbm: Stochastic Blockmodels}
\description{
A collection of tools and functions to adjust a variety of stochastic blockmodels (SBM). Supports at the moment Simple, Bipartite, 'Multipartite' and Multiplex SBM (undirected or directed with Bernoulli, Poisson or Gaussian emission laws on the edges, and possibly covariate for Simple and Bipartite SBM). See Léger (2016) \doi{10.48550/arXiv.1602.07587}, 'Barbillon et al.' (2020) \doi{10.1111/rssa.12193} and 'Bar-Hen et al.' (2020) \doi{10.48550/arXiv.1807.10138}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://grosssbm.github.io/sbm/}
  \item Report bugs at \url{https://github.com/GrossSBM/sbm/issues}
}

}
\author{
\strong{Maintainer}: Julien Chiquet \email{julien.chiquet@inrae.fr} (\href{https://orcid.org/0000-0002-3629-3429}{ORCID})

Authors:
\itemize{
  \item Julien Chiquet \email{julien.chiquet@inrae.fr} (\href{https://orcid.org/0000-0002-3629-3429}{ORCID})
  \item Sophie Donnet \email{sophie.donnet@inrae.fr} (\href{https://orcid.org/0000-0003-4370-7316}{ORCID})
  \item Pierre Barbillon \email{pierre.barbillon@agroparistech.fr} (\href{https://orcid.org/0000-0002-7766-7693}{ORCID})
}

Other contributors:
\itemize{
  \item großBM team [contributor]
}

}
\keyword{internal}
