% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-SBM_fit.R
\name{SBM_fit}
\alias{SBM_fit}
\title{R6 virtual class for SBM fit (mother class of Simple and Bipartite SBM fit)}
\description{
R6 virtual class for SBM fit (mother class of Simple and Bipartite SBM fit)

R6 virtual class for SBM fit (mother class of Simple and Bipartite SBM fit)
}
\section{Super class}{
\code{\link[GSBM:SBM]{GSBM::SBM}} -> \code{SBM_fit}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{probMemberships}}{matrix -- or list of 2 matrices for Bipartite network -- of estimated probabilities for block memberships for all nodes}

\item{\code{loglik}}{double: approximation of the log-likelihood (variational lower bound) reached}

\item{\code{ICL}}{double: value of the integrated classification loglielihood}

\item{\code{fitted}}{matrix of predicted value of the network}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SBM_fit$new()}}
\item \href{#method-clone}{\code{SBM_fit$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor for SBM fit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM_fit$new(data, model, covarList)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{the data matrix of the network}

\item{\code{model}}{character describing the type of model}

\item{\code{covarList}}{optional list of matrices for covariates}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM_fit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
