% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-ScoreSBM.R
\name{ScoreSBM}
\alias{ScoreSBM}
\title{R6 virtual class for ScoreSBM representation}
\description{
R6 virtual class for ScoreSBM representation

R6 virtual class for ScoreSBM representation
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{dimension}}{size-2 vector: dimension of the network}

\item{\code{modelName}}{character, the family of model for the distribution of the edges}

\item{\code{blockProp}}{vector of block proportions (aka prior probabilities of each block)}

\item{\code{connectParam}}{parameters associated to the connectivity of the SBM, e.g. matrix of inter/inter block probabilities when model is Bernoulli}

\item{\code{emissionParam}}{emission parameters for the Scores.}

\item{\code{listScore}}{the matrix (adjacency or incidence) encoding the network}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScoreSBM$new()}}
\item \href{#method-plot}{\code{ScoreSBM$plot()}}
\item \href{#method-show}{\code{ScoreSBM$show()}}
\item \href{#method-print}{\code{ScoreSBM$print()}}
\item \href{#method-clone}{\code{ScoreSBM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor for SBM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreSBM$new(
  model = NA,
  nbScores = NA,
  dimension = NA,
  directed = NA,
  blockProp = NA,
  connectParam = NA,
  emissionParam = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{character describing the type of model}

\item{\code{nbScores}}{number of scores}

\item{\code{dimension}}{dimension of the network matrix (nb of nodes)}

\item{\code{directed}}{directed (true or false)}

\item{\code{blockProp}}{parameters for block proportions (vector of list of vectors)}

\item{\code{connectParam}}{matrix of parameters for connectivity}

\item{\code{emissionParam}}{parameters of the distribution of the Scores}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
basic matrix plot method for Noisy SBM object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreSBM$plot(
  type = c("data", "predictedNetwork"),
  ordered = TRUE,
  color = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{character for the type of plot: either 'data' (true connection) or 'predictedNetwork' (fitted connection). Default to 'data'.}

\item{\code{ordered}}{logical: should the rows and columns be reoredered accordigin to the clustering? Default to \code{TRUE}.}

\item{\code{color}}{logical. Adapt colormap to the clustering. Default to \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\if{latex}{\out{\hypertarget{method-show}{}}}
\subsection{Method \code{show()}}{
print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreSBM$show(type = "Score Stochastic Block Model")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{character to tune the displayed name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreSBM$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreSBM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
