% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-ScoreSBM_fit.R
\name{ScoreSBM_fit}
\alias{ScoreSBM_fit}
\title{R6 Class definition of an Score  SBM fit}
\description{
R6 Class definition of an Score  SBM fit

R6 Class definition of an Score  SBM fit
}
\details{
This class is designed to give a representation and adjust an SBM fitted with blockmodels from Score matrices.
}
\section{Super classes}{
\code{\link[GSBM:SBM]{GSBM::SBM}} -> \code{\link[GSBM:SBM_fit]{GSBM::SBM_fit}} -> \code{ScoreSBM_fit}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{nbNodes}}{number of nodes}

\item{\code{nbBlocks}}{number of blocks}

\item{\code{nbDyads}}{number of dyads (potential edges in the network)}

\item{\code{memberships}}{vector of clustering}

\item{\code{directed}}{is the network directed or not}

\item{\code{scores}}{is the list of the scores of each dyad in the network}

\item{\code{nbScores}}{is the number of scores observed for each dyad}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScoreSBM_fit$new()}}
\item \href{#method-optimize}{\code{ScoreSBM_fit$optimize()}}
\item \href{#method-predict}{\code{ScoreSBM_fit$predict()}}
\item \href{#method-show}{\code{ScoreSBM_fit$show()}}
\item \href{#method-clone}{\code{ScoreSBM_fit$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="GSBM" data-topic="SBM" data-id="plot">}\href{../../GSBM/html/SBM.html#method-plot}{\code{GSBM::SBM$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GSBM" data-topic="SBM" data-id="print">}\href{../../GSBM/html/SBM.html#method-print}{\code{GSBM::SBM$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor for a Score SBM fit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreSBM_fit$new(scores, directed)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scores}}{a list of square (score) matrices}

\item{\code{directed}}{logical, directed network or not. In not, \code{adjacencyMatrix} must be symmetric.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-optimize}{}}}
\subsection{Method \code{optimize()}}{
function to perform optimization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreSBM_fit$optimize(estimOptions = list(), monitoring = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{estimOptions}}{options for the estimation}

\item{\code{monitoring}}{monitoring Options}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
prediction under the currently estimated model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreSBM_fit$predict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a matrix of expected values for each score and dyad
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\if{latex}{\out{\hypertarget{method-show}{}}}
\subsection{Method \code{show()}}{
show method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreSBM_fit$show(type = "Fit of a Score  Stochastic Block Model")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{character used to specify the type of SBM}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreSBM_fit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
