% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-ScoreSBM_sampler.R
\name{ScoreSBM_sampler}
\alias{ScoreSBM_sampler}
\title{R6 class for Score SBM sampler}
\description{
R6 class for Score SBM sampler

R6 class for Score SBM sampler
}
\section{Super classes}{
\code{\link[GSBM:SBM]{GSBM::SBM}} -> \code{\link[GSBM:SBM_sampler]{GSBM::SBM_sampler}} -> \code{\link[GSBM:SimpleSBM_sampler]{GSBM::SimpleSBM_sampler}} -> \code{ScoreSBM_sampler}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{nbScores}}{number of scores}

\item{\code{emissionParam}}{parameters of the emission distribution of the scores}

\item{\code{scores}}{a list of length nbScores, each matrix correspondonding to a score for each dyad}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScoreSBM_sampler$new()}}
\item \href{#method-rScores}{\code{ScoreSBM_sampler$rScores()}}
\item \href{#method-show}{\code{ScoreSBM_sampler$show()}}
\item \href{#method-clone}{\code{ScoreSBM_sampler$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="GSBM" data-topic="SBM" data-id="plot">}\href{../../GSBM/html/SBM.html#method-plot}{\code{GSBM::SBM$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GSBM" data-topic="SBM" data-id="print">}\href{../../GSBM/html/SBM.html#method-print}{\code{GSBM::SBM$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GSBM" data-topic="SimpleSBM_sampler" data-id="rAdjacency">}\href{../../GSBM/html/SimpleSBM_sampler.html#method-rAdjacency}{\code{GSBM::SimpleSBM_sampler$rAdjacency()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GSBM" data-topic="SimpleSBM_sampler" data-id="rMemberships">}\href{../../GSBM/html/SimpleSBM_sampler.html#method-rMemberships}{\code{GSBM::SimpleSBM_sampler$rMemberships()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor for ScoreSBM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreSBM_sampler$new(
  nbNodes,
  directed,
  blockProp,
  connectParam,
  emissionParam,
  seed = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nbNodes}}{number of nodes in the network}

\item{\code{directed}}{logical, directed network or not.}

\item{\code{blockProp}}{parameters for block proportions (vector of list of vectors)}

\item{\code{connectParam}}{list of parameters for connectivity with a matrix of means 'mu' and an optional scalar for the variance 'sigma2'. The size of mu must match \code{blockProp} length}

\item{\code{emissionParam}}{parameters of the emission of the Noisy SBM. List of two terms : noEdgeParam and edgeParam. Each element is a list of mean (vector or size nbScores) and variance matrix (var: square matrix of size nbScores)}

\item{\code{seed}}{integer to set seed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rScores"></a>}}
\if{latex}{\out{\hypertarget{method-rScores}{}}}
\subsection{Method \code{rScores()}}{
a method to sample the score matrices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreSBM_sampler$rScores()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing (sampled Scores are stored in the current object)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\if{latex}{\out{\hypertarget{method-show}{}}}
\subsection{Method \code{show()}}{
show method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreSBM_sampler$show(type = "Sampler for a Score Stochastic Block Model")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{character used to specify the type of SBM}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreSBM_sampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
