% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_sbm_llik.R
\name{tdmm_n_param}
\alias{tdmm_n_param}
\title{Calculate number of parameters for mixed model}
\usage{
tdmm_n_param(N, K, Time)
}
\arguments{
\item{N}{number of nodes}

\item{K}{number of blocks}

\item{Time}{number of time slices}
}
\description{
Calculate number of parameters for mixed model
}
