% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_sbm_llik.R
\name{tdmm_sbm_llik}
\alias{tdmm_sbm_llik}
\title{Calculate log-likelihood of TDMM-SBM (time-dependent mixed-membership stochastic block model)}
\usage{
tdmm_sbm_llik(A, C, omega)
}
\arguments{
\item{A}{is a A (time) series of network data represented  as N x N x Time array.}

\item{C}{is a N x K matrix of mixed group membership whose columns sum to 1}

\item{omega}{is a Time x K x K array describing block-to-block traffic at each time period or Time-length list of K x K matrices.}
}
\description{
Calculate log-likelihood of TDMM-SBM (time-dependent mixed-membership stochastic block model)
}
\examples{
# Uses output from python implementation (slighly reformatted)
data("la_byhour_edgelist", "la_mixed_roles_2", "la_mixed_omega_2")
A = edgelist_to_adj(la_byhour_edgelist, as.array = TRUE)
tdmm_sbm_llik(A, la_mixed_roles_2, la_mixed_omega_2)

}
