% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-BipartiteSBM.R
\name{BipartiteSBM}
\alias{BipartiteSBM}
\title{R6 virtual class for SBM representation}
\description{
R6 virtual class for SBM representation

R6 virtual class for SBM representation
}
\section{Super class}{
\code{\link[sbm:SBM]{sbm::SBM}} -> \code{BipartiteSBM}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{variational}}{probabilities of nodes being in a block}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{variational}}{probabilities of nodes being in a block}

\item{\code{memberships}}{list of size 2: vector of memberships in row, in column.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BipartiteSBM$new()}}
\item \href{#method-clone}{\code{BipartiteSBM$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="SBM" data-id="plot">}\href{../../sbm/html/SBM.html#method-plot}{\code{sbm::SBM$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="SBM" data-id="print">}\href{../../sbm/html/SBM.html#method-print}{\code{sbm::SBM$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="SBM" data-id="show">}\href{../../sbm/html/SBM.html#method-show}{\code{sbm::SBM$show()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor for a Simple SBM fit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BipartiteSBM$new(
  incidenceMatrix,
  model,
  dimLabels = list(row = "rowLabel", col = "colLabel"),
  covarList = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{incidenceMatrix}}{incidence (weighted) matrix}

\item{\code{model}}{character (\code{'bernoulli'}, \code{'poisson'}, \code{'gaussian'})}

\item{\code{dimLabels}}{list of labels of each dimension (in row, in columns)}

\item{\code{covarList}}{and optional list of covariates, each of whom must have the same dimension as \code{incidenceMatrix}}

\item{\code{directed}}{logical, directed network or not. In not, \code{incidenceMatrix} must be symmetric.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BipartiteSBM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
