% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-BipartiteSBM.R
\name{BipartiteSBM_fit}
\alias{BipartiteSBM_fit}
\title{R6 Class definition of an Bipartite SBM fit}
\description{
This class is designed to give a representation and adjust an LBM fitted with blockmodels.
}
\section{Super classes}{
\code{\link[GSBM:SBM]{GSBM::SBM}} -> \code{\link[GSBM:SBM_fit]{GSBM::SBM_fit}} -> \code{BipartiteSBM_fit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BipartiteSBM_fit$new()}}
\item \href{#method-optimize}{\code{BipartiteSBM_fit$optimize()}}
\item \href{#method-clone}{\code{BipartiteSBM_fit$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BipartiteSBM_fit$new(incidenceMatrix, model, covarList = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-optimize}{}}}
\subsection{Method \code{optimize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BipartiteSBM_fit$optimize(
  verbosity = 3,
  plot = "",
  explorFactor = 1.5,
  nbBlocksRange = c(4, Inf),
  nbCores = parallel::detectCores()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BipartiteSBM_fit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
