% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineNetwork.R
\name{defineNetwork}
\alias{defineNetwork}
\title{Define a network}
\usage{
defineNetwork(
  netMat,
  model = "bernoulli",
  type,
  directed = !isSymmetric(netMat),
  dimLabels = list(row = "rowLabel", col = "colLabels"),
  covariates = list()
)
}
\arguments{
\item{type}{Type of the matrix, choice between 'simple' and 'bipartite'}

\item{dimLabels}{an optional list of labels for each dimension (in row, in column)}
}
\value{
a network ready to be part of a multipartite network
}
\description{
Define a network
}
\examples{
A <- matrix(rbinom(100,1,.2),10,10)
type <- "simple"
defineNetwork(A,"poisson",type,directed=FALSE,dimLabels=c("Actor","Actor"))
}
