% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimateMultipartiteSBM}
\alias{estimateMultipartiteSBM}
\title{Estimation for multipartite SBM}
\usage{
estimateMultipartiteSBM(ldefinedNet, estimOptions = list())
}
\arguments{
\item{ldefinedNet}{list of networks that were defined by defineNetwork function}

\item{estimOptions}{option for estimation}
}
\value{

}
\description{
Estimation for multipartite SBM
}
