% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-MultipartiteSBM.R
\name{MultipartiteSBM}
\alias{MultipartiteSBM}
\title{R6 Class definition of a Multipartite SBM}
\description{
R6 Class definition of a Multipartite SBM

R6 Class definition of a Multipartite SBM
}
\details{
R6 virtual class for Multipartite SBM representation
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{nbNetworks}}{: number of networks in the multipartite network}

\item{\code{listSBM}}{: list of SimpleSBMs or BipartiteSBMs}

\item{\code{archiMultipartite}}{: organization of the multipartite network}

\item{\code{dimLabels}}{: labels of the functional groups}

\item{\code{nbLabels}}{: number of Functional groups involved in the multipartite}

\item{\code{nbNodes}}{: number of Nodes in each FG,}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MultipartiteSBM$new()}}
\item \href{#method-plot}{\code{MultipartiteSBM$plot()}}
\item \href{#method-clone}{\code{MultipartiteSBM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor for Multipartite SBM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultipartiteSBM$new(listSBM, memberships = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{listSBM}}{list of SimpleSBM or BipartiteSBM}

\item{\code{memberships}}{list of memberships for each node in each function group.Default value is NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
plot Multipartite Matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultipartiteSBM$plot(normalizing = FALSE, ordered = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{normalizing}}{TRUE if the various matrices are renormalized. FALSE otherwise. Default value = FALSE}

\item{\code{ordered}}{TRUE is the matrices are plotted after reorganization with the blocks. Default value = TRUE}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultipartiteSBM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
