% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-MultipartiteSBM_fit.R
\name{MultipartiteSBM_fit}
\alias{MultipartiteSBM_fit}
\title{R6 Class definition of a Multipartite SBM fit}
\description{
R6 Class definition of a Multipartite SBM fit

R6 Class definition of a Multipartite SBM fit
}
\details{
This class is designed to give a representation and adjust a Multipartite SBM fitted with GREMLIN.
}
\section{Super class}{
\code{\link[sbm:MultipartiteSBM]{sbm::MultipartiteSBM}} -> \code{MultipartiteSBM_fit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MultipartiteSBM_fit$new()}}
\item \href{#method-optimize}{\code{MultipartiteSBM_fit$optimize()}}
\item \href{#method-getBM}{\code{MultipartiteSBM_fit$getBM()}}
\item \href{#method-predict}{\code{MultipartiteSBM_fit$predict()}}
\item \href{#method-clone}{\code{MultipartiteSBM_fit$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="MultipartiteSBM" data-id="plot">}\href{../../sbm/html/MultipartiteSBM.html#method-plot}{\code{sbm::MultipartiteSBM$plot()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor for Multipartite SBM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultipartiteSBM_fit$new(listSBM)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{listSBM}}{list of SBM object with}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-optimize}{}}}
\subsection{Method \code{optimize()}}{
estimation via GREMLIN
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultipartiteSBM_fit$optimize(currentOptions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{currentOptions}}{options for MultipartiteBM}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBM"></a>}}
\if{latex}{\out{\hypertarget{method-getBM}{}}}
\subsection{Method \code{getBM()}}{
getBM returns a given network from a Multipartite SBM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultipartiteSBM_fit$getBM(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{index of the asked network}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
prediction under the currently estimated model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultipartiteSBM_fit$predict()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covarList}}{a list of covariates. By default, we use the covariates with which the model was estimated}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list of matrices matrix of expected values for each dyad
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultipartiteSBM_fit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
