% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-MultiplexSBM_fit.R
\name{MultiplexSBM_fit}
\alias{MultiplexSBM_fit}
\title{R6 Class definition of a Multiplex SBM fit}
\description{
R6 Class definition of a Multiplex SBM fit

R6 Class definition of a Multiplex SBM fit
}
\details{
This class is designed to give a representation and adjust a Multiplex SBM fitted with GREMLIN.

The list of parameters \code{estimOptions} essentially tunes the optimization process and the variational EM algorithm, with the following parameters
\itemize{
\item{"nbCores"}{integer for number of cores used.  Default is 2}
\item{"verbosity"}{integer for verbosity (0, 1). Default is 1}
\item{"nbBlocksRange"}{List of length the number of functional groups, each element supplying the minimal and maximal number of blocks to be explored. The names of the list must be the names of the functional groups.  Default value is from 1 to 10)}
\item{"initBM"}{Boolean. True if using simple and bipartite SBM as initialisations. Default value  = TRUE}
\item{"maxiterVEM"}{Number of max. number of iterations in  the VEM. Default value  = 100}
\item{"maxiterVE"}{Number of max. number of iterations in  the VE. Default value  = 100}
}
}
\section{Super classes}{
\code{\link[sbm:MultipartiteSBM]{sbm::MultipartiteSBM}} -> \code{\link[sbm:MultiplexSBM]{sbm::MultiplexSBM}} -> \code{MultiplexSBM_fit}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{memberships}}{a list with the memberships in all the functional groups}

\item{\code{probMemberships}}{or list of nbFG matrices for of estimated probabilities for block memberships for all nodes}

\item{\code{nbBlocks}}{: vector with the number of blocks in each FG}

\item{\code{blockProp}}{: block proportions in each function group}

\item{\code{connectParam}}{: connection parameters in each network}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MultiplexSBM_fit$new()}}
\item \href{#method-optimize}{\code{MultiplexSBM_fit$optimize()}}
\item \href{#method-clone}{\code{MultiplexSBM_fit$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="MultipartiteSBM" data-id="plot">}\href{../../sbm/html/MultipartiteSBM.html#method-plot}{\code{sbm::MultipartiteSBM$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="MultipartiteSBM" data-id="print">}\href{../../sbm/html/MultipartiteSBM.html#method-print}{\code{sbm::MultipartiteSBM$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="MultipartiteSBM" data-id="show">}\href{../../sbm/html/MultipartiteSBM.html#method-show}{\code{sbm::MultipartiteSBM$show()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor for Multiplex SBM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiplexSBM_fit$new(listSBM, dep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{listSBM}}{list of SBM object with}

\item{\code{dep}}{boolean indicating whether dependence is assumed between networks beyond the common dependence on the latent variables}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-optimize}{}}}
\subsection{Method \code{optimize()}}{
estimation of multipartiteSBM via GREMLINS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiplexSBM_fit$optimize(estimOptions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{estimOptions}}{options for MultipartiteBM}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiplexSBM_fit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
