% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-MultipartiteSBM.R
\name{plot.MultipartiteSBM}
\alias{plot.MultipartiteSBM}
\title{MultipartiteSBM Plot}
\usage{
\method{plot}{MultipartiteSBM}(
  x,
  type = c("data", "expected", "meso"),
  ordered = TRUE,
  plotOptions = list(),
  ...
)
}
\arguments{
\item{x}{an object inheriting from class MultipartiteSBM}

\item{type}{character for the type of plot: either 'data' (true connection) or 'expected' (fitted connection) or 'meso' (meso-scopic). Default to 'data'.}

\item{ordered}{logical: should the functional group be ordered according to the clustering? Default to \code{TRUE}.}

\item{plotOptions}{list with parameters.}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
a ggplot2 object for the \code{'data'} and \code{'expected'}, a list with the igraph object \code{g}, the \code{layout} and the \code{plotOptions} for the \code{'meso'}
}
\description{
Basic matrix plot method for SBM object
}
\details{
The list of parameters \code{plotOptions} for the mesoscopic plot is
\itemize{
\item{"seed": }{seed to control the layout}
\item{"title": }{character string for the title. Default value is NULL}
\item{"layout": }{Default value = NULL}
\item{"vertex.color": }{Default value is "salmon2"}
\item{"vertex.frame.color": }{Node border color.Default value is "black" }
\item{"vertex.shape": }{One of "none", "circle", "square", "csquare", "rectangle" "crectangle", "vrectangle", "pie", "raster", or "sphere". Default value = "circle"}
\item{"vertex.size": }{Size of the nodes (default factor is 1). Vector of length the number of FG}
\item{"vertex.size2": }{The second size of the node (e.g. for a rectangle)}
\item{"vertex.label": }{Names of the vertices. Default value is the label of the nodes}
\item{"vertex.label.color": }{Default value is  "black"}
\item{"vertex.label.font": }{Default value is 2. Font: 1 plain, 2 bold, 3, italic, 4 bold italic, 5 symbol}
\item{"vertex.label.cex": }{Font size (multiplication factor, device-dependent).Default value is  0.9.}
\item{"vertex.label.dist": }{Distance between the label and the vertex. Default value is  0}
\item{"vertex.label.degree": }{The position of the label in relation to the vertex. default value is 0}
\item{"edge.threshold": }{Threshold under which the edge is not plotted. Default value is = -Inf}
\item{"edge.color": }{Default value is "gray"}
\item{"edge.width": }{Factor parameter. Default value is 1}
\item{"edge.arrow.size": }{Default value is 1}
\item{"edge.arrow.width": }{Default value is 2}
\item{"edge.lty": }{Line type, could be 0 or "blank", 1 or "solid", 2 or "dashed", 3 or "dotted", 4 or "dotdash", 5 or "longdash", 6 or "twodash". Default value is "solid"}
\item{"edge.curved": }{Default value is = 0.3}
}
The list of parameters \code{plotOptions} for the matrix plot is
\itemize{
\item{"normalized":}{Boolean. TRUE if the various matrices are presented in the same scale (between O and 1). FALSE otherwise. Default value FALSE}
\item{"compact":}{Boolean. Default value is TRUE if you ask for the matrices to be transposed to have a more compact view}
\item{"legend": }{Boolean. Set TRUE if you   want to see the legend. Default value is FALSE}
\item{"legend.title": }{Boolean. Set TRUE if you want to print the title of the legend. Default value is FALSE}
\item{"legend.position": }{Position of the legend. Possible values are 'bottom', 'top','left,'right'. Default value is 'bottom'}
\item{"nodeNames": }{Set true if the node Names must be plotted. Default value is FALSE}
\item{"line.color":}{The color of the lines to separate groups. Default value is red}
\item{"line.width":}{Width  of the lines to separate groups. Default value is NULL, automatically chosen}
\item{"title": }{Title of the plot. Default value is NULL}
}
}
