% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-SBM.R
\name{plot.SBM}
\alias{plot.SBM}
\title{SBM Plot}
\usage{
\method{plot}{SBM}(x, type = c("data", "expected"), ordered = TRUE, color = TRUE, ...)
}
\arguments{
\item{x}{an R6 object inheriting from class SBM_fit (like SimpleSBM_fit or BipartiteSBM_fit)}

\item{type}{character for the type of plot: either 'data' (true connection) or 'expected' (fitted connection). Default to 'data'.}

\item{ordered}{logical: should the rows and columns be reordered according to the clustering? Default to \code{TRUE}.}

\item{color}{logical. Adapt colormap to the clustering. Default to \code{TRUE}.}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\description{
Basic matrix plot method for SBM object
}
