% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-SBM.R
\name{plot.SBM}
\alias{plot.SBM}
\title{SBM Plot}
\usage{
\method{plot}{SBM}(
  x,
  type = c("data", "expected"),
  ordered = TRUE,
  rowLabel = NULL,
  colLabel = NULL,
  fileNameSave = NULL,
  ...
)
}
\arguments{
\item{type}{character for the type of plot: either 'data' (true connection) or 'expected' (fitted connection). Default to 'data'.}

\item{ordered}{logical: should the rows and columns be reoredered according to the clustering? Default to \code{TRUE}.}

\item{rowLabel}{character : type of the individual in row. Default to \code{NULL}.}

\item{colLabel}{character : type of the individual in col. Default to \code{NULL}.}

\item{fileNameSave}{character : name of the file to save the plot Default to \code{NULL}.}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\description{
basic matrix plot method for SBM object
}
\details{
Basic matrix plot method for SBM object
}
