% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMyMatrix.R
\name{plotMyMatrix}
\alias{plotMyMatrix}
\title{Plot an adjacency or incidence Matrix}
\usage{
plotMyMatrix(Mat, rowLabel = NULL, colLabel = NULL)
}
\arguments{
\item{Mat}{: a matrix representing the network}

\item{rowLabel}{character : type of nodes in rows (functional group) (Default is \code{NULL})}

\item{colLabel}{character :  type of nodes in columns (functional group) (Default is \code{NULL})}
}
\value{
a ggplot object corresponding to the plot
}
\description{
Plot an adjacency or incidence Matrix
}
\examples{
M <- matrix(sample(c(0,1),900,replace=TRUE),30,30)
plotMyMatrix(M)
M2 <- matrix(rpois(800,10),40,20)
plotMyMatrix(M2,rowLabel='ind',colLabel = 'book')

}
