% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMyMatrix.R
\name{plotMyMatrix}
\alias{plotMyMatrix}
\title{Plot an adjacency or incidence Matrix}
\usage{
plotMyMatrix(Mat, dimLabels = list(row = NULL, col = NULL))
}
\arguments{
\item{Mat}{: a matrix representing the network}

\item{dimLabels}{: a list of length 2 specifying the types of nodes in row and col  (functional group) (Default is \code{NULL})}
}
\value{
a ggplot object corresponding to the plot
}
\description{
Plot an adjacency or incidence Matrix
}
\examples{
M <- matrix(sample(c(0,1),900,replace=TRUE),30,30)
plotMyMatrix(M)
M2 <- matrix(rpois(800,10),40,20)
plotMyMatrix(M2,dimLabels=list(row= 'reader',col = 'book'))

}
