% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMyMultipartite.R
\name{plotMyMultipartiteMatrix}
\alias{plotMyMultipartiteMatrix}
\title{Plot the matrices corresponding to a Multipartite Network}
\usage{
plotMyMultipartiteMatrix(listSBM, normalizing = FALSE, memberships = NULL)
}
\arguments{
\item{listSBM}{: a list of objects representing the multipartite network (see)}

\item{memberships}{: a list of length number of Functional Groups with clusterings inside these FG}
}
\value{
a ggplot object corresponding to the plot
}
\description{
Plot the matrices corresponding to a Multipartite Network
}
\examples{
data(MPEcoNetwork, package = "GREMLIN")
PlantFlovis = defineSBM(MPEcoNetwork$Inc_plant_flovis, type = "bipartite",model = 'bernoulli',dimLabels = list(row="Plants",col="Flovis"))
PlantAnt = defineSBM(MPEcoNetwork$Inc_plant_ant,type = "bipartite",model = 'bernoulli',dimLabels =list(row = "Plants", col = "Ants"))
PlantBird = defineSBM(MPEcoNetwork$Inc_plant_bird,type = "bipartite",model = 'bernoulli',dimLabels =list(row = "Plants",col = "Birds"))
plotMyMultipartiteMatrix(list(PlantFlovis,PlantAnt,PlantBird))


}
