% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMyMultipartite.R
\name{plotMyMultipartiteMatrix}
\alias{plotMyMultipartiteMatrix}
\title{Plot the matrices corresonding to a Multipartite Network}
\usage{
plotMyMultipartiteMatrix(list_Net)
}
\arguments{
\item{list_Net}{: a list of objects representing the multipartite network (see)}
}
\value{
a ggplot object corresponding to the plot
}
\description{
Plot the matrices corresonding to a Multipartite Network
}
\examples{
data(MPEcoNetwork, package = "GREMLIN")
PlantFlovis = sbm::defineNetwork(MPEcoNetwork$Inc_plant_flovis, type = "bipartite",model = 'bernoulli',dimLabels = list(row="Plants",col="Flovis"))
PlantAnt = sbm::defineNetwork(MPEcoNetwork$Inc_plant_ant,type = "bipartite",model = 'bernoulli',dimLabels =list(row = "Plants", col = "Ants"))
PlantBird = sbm::defineNetwork(MPEcoNetwork$Inc_plant_bird,type = "bipartite",model = 'bernoulli',dimLabels =list(row = "Plants",col = "Birds"))
plotMyMultipartiteMatrix(list(PlantFlovis,PlantAnt,PlantBird)


}
