% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sampleBipartiteSBM}
\alias{sampleBipartiteSBM}
\title{Sampling of Bipartite SBMs}
\usage{
sampleBipartiteSBM(
  nbNodes,
  blockProp,
  connectParam,
  model = "bernoulli",
  covariates = list(),
  covariatesParam = numeric(0)
)
}
\arguments{
\item{nbNodes}{number of nodes in the network}

\item{blockProp}{parameters for block proportions: list of size two with row and column block proportions}

\item{connectParam}{list of parameters for connectivity with a matrix of means 'mu' and an optional matrix of variances 'sigma2', the sizes of which must match \code{blockProp} length (in row, respectively in column)}

\item{model}{character describing the model for the relation between nodes (\code{'bernoulli'}, \code{'poisson'}, \code{'gaussian'}, ...). Default is \code{'bernoulli'}.}

\item{covariates}{a list of matrices with same dimension as mat describing covariates at the edge level. No covariate per Default.}

\item{covariatesParam}{optional vector of covariates effect. A zero length numeric vector by default.}
}
\value{
an object with class BipartiteSBM_sampler
}
\description{
This function samples a simple Stochastic Block Models, with various model
for the distribution of the edges:  Bernoulli, Poisson, or Gaussian models, and possibly with covariates
}
\examples{
### BIPARTITE SBM
## Graph parameters
nbNodes <- c(100, 120)
blockProp <- list(c(.5, .5), c(1/3,1/3,1/3)) # group proportions
connectProb <- matrix(runif(6), 2, 3)

## Graph Sampling
mySampler <- sampleBipartiteSBM(nbNodes, blockProp, list(mu = connectProb))
plot(mySampler)
mySampler$rMemberships() # sample new memberships
mySampler$rIncidence()   # sample new incidence matrix
plot(mySampler)
}
