% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sampleSimpleSBM}
\alias{sampleSimpleSBM}
\title{Sampling of Simple SBMs}
\usage{
sampleSimpleSBM(
  nbNodes,
  blockProp,
  connectParam,
  model = "bernoulli",
  directed = FALSE,
  covariates = list(),
  covariatesParam = numeric(0)
)
}
\arguments{
\item{nbNodes}{number of nodes in the network}

\item{blockProp}{parameters for block proportions}

\item{connectParam}{list of parameters for connectivity with a matrix of means 'mu' and an optional matrix of variances 'sigma2', the sizes of which must match \code{blockProp} length}

\item{model}{character describing the model for the relation between nodes (\code{'bernoulli'}, \code{'poisson'}, \code{'gaussian'}, ...). Default is \code{'bernoulli'}.}

\item{directed}{logical, directed network or not. Default is \code{FALSE}.}

\item{covariates}{a list of matrices with same dimension as mat describing covariates at the edge level. No covariate per Default.}

\item{covariatesParam}{optional vector of covariates effect. A zero length numeric vector by default.}
}
\value{
an object with class SimpleSBM_sampler
}
\description{
This function samples a simple Stochastic Block Models, with various model
for the distribution of the edges:  Bernoulli, Poisson, or Gaussian models, and possibly with covariates
}
\examples{
### SIMPLE SBM - Bernoulli
nbNodes  <- 90
blockProp <- c(.5, .25, .25)      # group proportions
connectProb <- diag(.4, 3) + 0.05 # connectivity matrix: affiliation network

## Graph Sampling
mySampler <- sampleSimpleSBM(nbNodes, blockProp, list(mu = connectProb))
plot(mySampler)
mySampler$rBlocks()     # sample new blocks
mySampler$rAdjMatrix() # sample new adjacency matrix
plot(mySampler)
}
