% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm-package.R
\docType{package}
\name{sbm-package}
\alias{sbm}
\alias{sbm-package}
\title{sbm: Stochastic Blockmodels}
\description{
A collection of tools and function to adjust a variety of stochastic blockmodels (SBM), 
  such as Simple (directed/undirected) SBM, Bipartite SBM, Multipartite SBM, Multilevel SBM and more.
}
\author{
\strong{Maintainer}: Julien Chiquet \email{julien.chiquet@inrae.fr} (\href{https://orcid.org/0000-0002-3629-3429}{ORCID})

Authors:
\itemize{
  \item Sophie Donnet \email{sophie.donnet@inrae.fr}
}

Other contributors:
\itemize{
  \item Pierre Barbillon \email{pierre.barbillon@agroparistech.fr} (\href{https://orcid.org/0000-0002-7766-7693}{ORCID}) [contributor]
}

}
\keyword{internal}
