% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm-package.R
\docType{package}
\name{sbm-package}
\alias{sbm}
\alias{sbm-package}
\title{sbm: Stochastic Blockmodels}
\description{
A collection of tools and function to adjust a variety of stochastic blockmodels (SBM). 
  Support at the moment Simple and Bipartite SBM (undirected and directed) for Bernoulli, Poisson and 
  Gaussian emission laws of the edges as described in Léger, 2016 <arxiv:1602.07587>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://grosssbm.github.io/sbm/}
  \item Report bugs at \url{https://github.com/GrossSBM/sbm/issues}
}

}
\author{
\strong{Maintainer}: Julien Chiquet \email{julien.chiquet@inrae.fr} (\href{https://orcid.org/0000-0002-3629-3429}{ORCID})

Authors:
\itemize{
  \item Sophie Donnet \email{sophie.donnet@inrae.fr}
  \item Pierre Barbillon \email{pierre.barbillon@agroparistech.fr} (\href{https://orcid.org/0000-0002-7766-7693}{ORCID})
}

Other contributors:
\itemize{
  \item großBM team [contributor]
}

}
\keyword{internal}
