% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-SBM.R
\name{SBM}
\alias{SBM}
\title{R6 virtual class for SBM representation (mother class of Simple and Bipartite SBM fit and sampler)}
\description{
R6 virtual class for SBM representation (mother class of Simple and Bipartite SBM fit and sampler)

R6 virtual class for SBM representation (mother class of Simple and Bipartite SBM fit and sampler)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{character, }}{the family of model for the distribution of the edges}

\item{\code{integer, }}{the number of covariates}

\item{\code{vector}}{of block proportions (aka prior probabilities of each block)}

\item{\code{parameters}}{associated to the connectivity of the SBM, e.g. matrix of inter/inter block probabilities when model is Bernoulli}

\item{\code{vector}}{of regression parameters assocaited with the covariates.}

\item{\code{list}}{of matrices of covariates}

\item{\code{effect}}{of covariates}

\item{\code{the}}{matrix (adjacency or incidence) encoding the network}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{character, }}{the family of model for the distribution of the edges}

\item{\code{integer, }}{the number of covariates}

\item{\code{vector}}{of block proportions (aka prior probabilities of each block)}

\item{\code{parameters}}{associated to the connectivity of the SBM, e.g. matrix of inter/inter block probabilities when model is Bernoulli}

\item{\code{vector}}{of regression parameters assocaited with the covariates.}

\item{\code{list}}{of matrices of covariates}

\item{\code{effect}}{of covariates}

\item{\code{the}}{matrix (adjacency or incidence) encoding the network}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SBM$new()}}
\item \href{#method-clone}{\code{SBM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor for SBM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM$new(
  model = NA,
  dimension = NA,
  blockProp = NA,
  connectParam = NA,
  covarParam = numeric(0),
  covarList = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
