% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-SBM.R
\name{SBM}
\alias{SBM}
\title{R6 virtual class for SBM representation (mother class of Simple and Bipartite SBM fit and sampler)}
\description{
R6 virtual class for SBM representation (mother class of Simple and Bipartite SBM fit and sampler)

R6 virtual class for SBM representation (mother class of Simple and Bipartite SBM fit and sampler)
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{dimension}}{size-2 vector: dimension of the network}

\item{\code{modelName}}{character, the family of model for the distribution of the edges}

\item{\code{directed}}{mode of the network data (directed or not or not applicable)}

\item{\code{dimLabels}}{vector or list of characters, the label of each dimension}

\item{\code{nbCovariates}}{integer, the number of covariates}

\item{\code{blockProp}}{block proportions (aka prior probabilities of each block)}

\item{\code{connectParam}}{parameters associated to the connectivity of the SBM, e.g. matrix of inter/inter block probabilities when model is Bernoulli}

\item{\code{covarParam}}{vector of regression parameters associated with the covariates.}

\item{\code{covarList}}{list of matrices of covariates}

\item{\code{covarArray}}{the array of covariates}

\item{\code{covarEffect}}{effect of covariates}

\item{\code{networkData}}{the network data (adjacency or incidence matrix or list of such object)}

\item{\code{expectation}}{expected values of connection under the currently adjusted model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SBM$new()}}
\item \href{#method-plot}{\code{SBM$plot()}}
\item \href{#method-show}{\code{SBM$show()}}
\item \href{#method-print}{\code{SBM$print()}}
\item \href{#method-clone}{\code{SBM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor for SBM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM$new(
  model = vector("character", 0),
  directed = vector("logical", 0),
  dimension = vector("numeric", 0),
  dimLabels = vector("character", 0),
  blockProp = vector("numeric", 0),
  connectParam = vector("list", 0),
  covarParam = numeric(length(covarList)),
  covarList = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{character describing the type of model}

\item{\code{directed}}{logical describing if the network data is directed or not}

\item{\code{dimension}}{dimension of the network data}

\item{\code{dimLabels}}{labels of each dimension}

\item{\code{blockProp}}{parameters for block proportions (vector or list of vectors)}

\item{\code{connectParam}}{list of parameters for connectivity}

\item{\code{covarParam}}{optional vector of covariates effect}

\item{\code{covarList}}{optional list of covariates data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
basic matrix plot method for SBM object or mesoscopic plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM$plot(
  type = c("data", "expected", "meso"),
  ordered = TRUE,
  plotOptions = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{character for the type of plot: either 'data' (true connection), 'expected' (fitted connection) or 'meso' (mesoscopic view). Default to 'data'.}

\item{\code{ordered}}{logical: should the rows and columns be reordered according to the clustering? Default to \code{TRUE}.}

\item{\code{plotOptions}}{list with the parameters for the plot}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The list of parameters \code{plotOptions}  for the mesoscopic plot is:
\itemize{
\item{"seed": }{seed to control the layout}
\item{"title": }{character string for the title. Default value is NULL}
\item{"layout": }{Default value = NULL}
\item{"vertex.color": }{Default value is "salmon2"}
\item{"vertex.frame.color": }{Node border color.Default value is "black" }
\item{"vertex.shape": }{One of "none", "circle", "square", "csquare", "rectangle" "crectangle", "vrectangle", "pie", "raster", or "sphere". Default value = "circle"}
\item{"vertex.size": }{Size of the node (default is 2)}
\item{"vertex.size2": }{The second size of the node (e.g. for a rectangle)}
\item{"vertex.label.name": }{Names of the vertices. Default value is the label of the nodes}
\item{"vertex.label.color": }{Default value is  "black"}
\item{"vertex.label.font": }{Default value is 2. Font: 1 plain, 2 bold, 3, italic, 4 bold italic, 5 symbol}
\item{"vertex.label.cex": }{Font size (multiplication factor, device-dependent).Default value is  0.9.}
\item{"vertex.label.dist": }{Distance between the label and the vertex. Default value is  0}
\item{"vertex.label.degree": }{The position of the label in relation to the vertex. default value is 0}
\item{"edge.threshold": }{Threshold under which the edge is not plotted. Default value is = -Inf}
\item{"edge.color": }{Default value is "gray"}
\item{"edge.width": }{Factor parameter. Default value is 10}
\item{"edge.arrow.size": }{Default value is 1}
\item{"edge.arrow.width": }{Default value is 2}
\item{"edge.lty": }{Line type, could be 0 or "blank", 1 or "solid", 2 or "dashed", 3 or "dotted", 4 or "dotdash", 5 or "longdash", 6 or "twodash". Default value is "solid"}
\item{"edge.curved": }{Default value is = 0.3.}
}
For type = 'data' or 'expected plot', the list of parameters \code{plotOptions} is
\itemize{
\item{"legend": }{Boolean. Set TRUE if you want to see the legend. Default value is FALSE}
\item{"legend.title":}{Boolean. Set TRUE if you want to print the title of the legend. Default value is FALSE}
\item{"legend.position":}{Position of the legend. Possible values are 'bottom', 'top','left,'right'. Default value is 'bottom'}
\item{"rowNames":}{Set true if the rownames must be plotted. Default value is FALSE}
\item{"colNames":}{Set true if the colNames must be plotted. Default value is FALSE}
\item{"line.color": }{Chain of character. The color of the lines to separate groups if a clustering is provided. Default value is red}
\item{"line.width": }{Numeric. Width  of the lines to separate groups. Default value is NULL, automatically chosen}
\item{"title": }{Chain of character. Title of the plot. Default value is NULL}
}
}

\subsection{Returns}{
a ggplot2 object for the \code{'data'} and \code{'expected'}, a list with the igraph object \code{g}, the \code{layout} and the \code{plotOptions} for the \code{'meso'}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\if{latex}{\out{\hypertarget{method-show}{}}}
\subsection{Method \code{show()}}{
print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM$show(type = "Stochastic Block Model")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{character to tune the displayed name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
